Contributing (#20)
* Create CONTRIBUTING.md

Started writing the contributing.md file based on GitHub's suggestions.

* Update CONTRIBUTING.md

Add a section on how to submit a (good) feature request.

* Update CONTRIBUTING.md

Forgot to replace the name. (Kudos to the Atom team for their nice CONTRIBUTING.md).

* Update CONTRIBUTING.md

It's a request not a (feature).

* Update CONTRIBUTING.md

Grammar!

* Update CONTRIBUTING.md

Add a feature request template for new exporters.

* Update CONTRIBUTING.md

Add the missing character.

* Update CONTRIBUTING.md

Make sure we find the feature requests!