String utils (#45)
* Update string_utils.c

The "string_recat" function now checks if the destination is NULL and uses "strdup" if so. This allows to initialize the description field to NULL.

* Update export_md.c

Check if the field contains information by comparing the strings to NULL. Output them only if they contain informations.

* Update object.c

Initialize the description field to NULL.
Check if the description field contains informations by comparing it to NULL. Only free it if it contains informations.